feat: export the type of loadConfigFile
