+++ rollup.config.ts
@@ -12,7 +12,7 @@ import addCliEntry from './build-plugins/add-cli-entry';
 import { moduleAliases } from './build-plugins/aliases';
 import cleanBeforeWrite from './build-plugins/clean-before-write';
 import conditionalFsEventsImport from './build-plugins/conditional-fsevents-import';
-import copyTypes from './build-plugins/copy-types';
+import copyTypes, { copyAllTypes } from './build-plugins/copy-types';
 import emitModulePackageFile from './build-plugins/emit-module-package-file';
 import esmDynamicImport from './build-plugins/esm-dynamic-import';
 import getLicenseHandler from './build-plugins/generate-license-file';
@@ -83,7 +83,7 @@ export default async function (
 			addCliEntry(),
 			esmDynamicImport(),
 			!command.configTest && collectLicenses(),
-			!command.configTest && copyTypes('rollup.d.ts')
+			!command.configTest && copyAllTypes()
 		],
 		strictDeprecations: true,
 		treeshake
